submissions/practicum1_groupname/backend_similarity/requirements.txt
pandas == 1.1.0
numpy == 1.19.1
requests == 2.24.0
Flask == 1.1.2
dask[dataframe] == 2.28.0
dask-image == 0.4.0
dask-distance == 0.2.0
dask[array] == 2.28.0
joblib == 0.16.0
pillow == 7.2.0
gcsfs==0.7.1
scikit-learn==0.23.2
